Refactor app routes and utilities
