+++ app.py
@@ -49,6 +49,7 @@ from retrorecon import (
     import_utils,
     status as status_mod,
     app_utils,
+    middleware,
 )
 from retrorecon.filters import manifest_links, oci_obj
 
@@ -59,16 +60,6 @@ app.add_template_filter(manifest_links, name="manifest_links")
 app.add_template_filter(oci_obj, name="oci_obj")
 
 
-@app.route('/favicon.ico', endpoint='core.favicon_ico')
-def favicon_ico() -> Response:
-    """Serve the favicon.ico from the application root."""
-    return send_from_directory(app.root_path, 'favicon.ico', mimetype='image/vnd.microsoft.icon')
-
-
-@app.route('/favicon.svg')
-def favicon_svg() -> Response:
-    """Serve the favicon.svg from the application root."""
-    return send_from_directory(app.root_path, 'favicon.svg', mimetype='image/svg+xml')
 
 def get_db_folder() -> str:
     """Return the folder where database files are stored."""
@@ -109,352 +100,12 @@ TEMP_DB_NAME = app_utils.TEMP_DB_NAME
 TEMP_DISPLAY_NAME = app_utils.TEMP_DISPLAY_NAME
 
 
-def _create_temp_db() -> None:
-    """Create a fresh temporary database for this session."""
-    app.config['DATABASE'] = os.path.join(app_utils.get_db_folder(), TEMP_DB_NAME)
-    if os.path.exists(app.config['DATABASE']):
-        os.remove(app.config['DATABASE'])
-    init_db()
-
 
 if not env_db:
     with app.app_context():
-        _create_temp_db()
-
-def _db_loaded() -> bool:
-    """Return True if a database file is currently configured and exists."""
-
-    return app_utils._db_loaded()
-
-def set_import_progress(status: str, message: str = '', current: int = 0, total: int = 0) -> None:
-    app_utils.set_import_progress(IMPORT_PROGRESS_FILE, status, message, current, total)
+        app_utils._create_temp_db()
 
 
-def get_import_progress() -> Dict[str, Any]:
-    return app_utils.get_import_progress(IMPORT_PROGRESS_FILE)
-
-
-def clear_import_progress() -> None:
-    app_utils.clear_import_progress(IMPORT_PROGRESS_FILE)
-
-
-
-def export_url_data(ids: Optional[List[int]] = None, query: str = '') -> List[Dict[str, Any]]:
-    """Return URL records filtered by ids or search query."""
-    return app_utils.export_url_data(ids=ids, query=query)
-
-
-
-
-def delete_subdomain(root_domain: str, subdomain: str) -> None:
-    """Remove a subdomain entry from the database."""
-    app_utils.delete_subdomain(root_domain, subdomain)
-
-
-@app.route('/', methods=['GET'])
-def index() -> str:
-    """Render the main search page or serve registry explorer views."""
-    repo_param = request.args.get("repo")
-    image_param = request.args.get("image")
-    if repo_param:
-        from retrorecon.routes.oci import repo_view
-        return repo_view(repo_param)
-    if image_param:
-        from retrorecon.routes.oci import image_view
-        return image_view(image_param)
-
-    q = request.args.get('q', '').strip()
-    is_tool_route = request.path in [
-        '/tools/jwt',
-        '/tools/screenshotter',
-        '/tools/subdomonster',
-        '/tools/text_tools'
-    ]
-    if not any([q, repo_param, image_param, request.args.get('tool'), is_tool_route]):
-        from retrorecon.routes.overview import _collect_counts, _collect_domains
-        if _db_loaded():
-            actual_name = os.path.basename(app.config['DATABASE'])
-            if actual_name == TEMP_DB_NAME:
-                actual_name = TEMP_DISPLAY_NAME
-        else:
-            actual_name = '(none)'
-        if session.get('db_display_name') != actual_name:
-            session['db_display_name'] = actual_name
-        db_name = session['db_display_name']
-        default_theme = 'nostalgia.css' if 'nostalgia.css' in AVAILABLE_THEMES else (AVAILABLE_THEMES[0] if AVAILABLE_THEMES else '')
-        current_theme = session.get('theme', default_theme)
-        data = {
-            'db_name': db_name,
-            'counts': _collect_counts(),
-            'domains': _collect_domains(),
-            'current_theme': current_theme,
-        }
-        return render_template('overview.html', **data)
-
-    select_all_matching = request.args.get('select_all_matching', 'false').lower() == 'true'
-    try:
-        page = int(request.args.get('page', 1))
-    except ValueError:
-        page = 1
-
-    tool = request.args.get('tool')
-    if request.path == '/tools/jwt':
-        tool = 'jwt'
-    elif request.path == '/tools/screenshotter':
-        tool = 'screenshot'
-    elif request.path == '/tools/subdomonster':
-        tool = 'subdomonster'
-    elif request.path == '/tools/text_tools':
-        tool = 'text'
-
-    sort = request.args.get('sort', 'id')
-    direction = request.args.get('dir', 'desc').lower()
-    if direction not in ['asc', 'desc']:
-        direction = 'desc'
-
-    try:
-        items_per_page = int(session.get('items_per_page', ITEMS_PER_PAGE))
-    except ValueError:
-        items_per_page = ITEMS_PER_PAGE
-    if items_per_page not in ITEMS_PER_PAGE_OPTIONS:
-        items_per_page = ITEMS_PER_PAGE
-
-    if _db_loaded():
-        where_clauses = []
-        params = []
-        if q:
-            try:
-                search_sql, search_params = search_utils.build_search_sql(q)
-                where_clauses.append(search_sql)
-                params.extend(search_params)
-            except Exception:
-                where_clauses.append(
-                    "("
-                    "url LIKE ? OR tags LIKE ? OR "
-                    "CAST(timestamp AS TEXT) LIKE ? OR "
-                    "CAST(status_code AS TEXT) LIKE ? OR "
-                    "mime_type LIKE ?"
-                    ")"
-                )
-                params.extend([f"%{q}%"] * 5)
-        where_sql = ""
-        if where_clauses:
-            where_sql = "WHERE " + " AND ".join(where_clauses)
-
-        sort_map = {
-            'url': 'url',
-            'timestamp': 'timestamp',
-            'status_code': 'status_code',
-            'mime_type': 'mime_type',
-            'id': 'id'
-        }
-        sort_col = sort_map.get(sort, 'id')
-
-        count_sql = f"SELECT COUNT(*) AS cnt FROM urls {where_sql}"
-        count_row = query_db(count_sql, params, one=True)
-        total_count = count_row['cnt'] if count_row else 0
-
-        total_pages = max(1, (total_count + items_per_page - 1) // items_per_page)
-
-        if page < 1:
-            page = 1
-        elif page > total_pages:
-            page = total_pages
-
-        offset = (page - 1) * items_per_page
-        select_sql = f"""
-            SELECT id, url, timestamp, status_code, mime_type, tags
-            FROM urls
-            {where_sql}
-            ORDER BY {sort_col} {direction.upper()}
-            LIMIT ? OFFSET ?
-        """
-        rows = query_db(select_sql, params + [items_per_page, offset])
-    else:
-        rows = []
-        total_pages = 1
-        total_count = 0
-
-    if _db_loaded():
-        actual_name = os.path.basename(app.config['DATABASE'])
-        if actual_name == TEMP_DB_NAME:
-            actual_name = TEMP_DISPLAY_NAME
-    else:
-        actual_name = '(none)'
-    if session.get('db_display_name') != actual_name:
-        session['db_display_name'] = actual_name
-    db_name = session['db_display_name']
-
-    try:
-        saved_dbs = sorted([
-            f for f in os.listdir(app_utils.get_db_folder())
-            if f.endswith('.db') and os.path.isfile(os.path.join(app_utils.get_db_folder(), f))
-        ])
-    except OSError:
-        saved_dbs = []
-
-    default_theme = 'nostalgia.css' if 'nostalgia.css' in AVAILABLE_THEMES else (AVAILABLE_THEMES[0] if AVAILABLE_THEMES else '')
-    current_theme = session.get('theme', default_theme)
-
-    default_background = 'background.jpg' if 'background.jpg' in AVAILABLE_BACKGROUNDS else (AVAILABLE_BACKGROUNDS[0] if AVAILABLE_BACKGROUNDS else '')
-    current_background = session.get('background', default_background)
-
-    panel_opacity = float(session.get('panel_opacity', 0.75))
-
-    search_history = session.get('search_history', [])
-    if q:
-        if q in search_history:
-            search_history.remove(q)
-        search_history.insert(0, q)
-        session['search_history'] = search_history[:10]
-
-    return render_template(
-        'index.html',
-        urls=rows,
-        page=page,
-        total_pages=total_pages,
-        q=q,
-        themes=AVAILABLE_THEMES,
-        theme_swatches=THEME_SWATCHES,
-        current_theme=current_theme,
-        backgrounds=AVAILABLE_BACKGROUNDS,
-        current_background=current_background,
-        panel_opacity=panel_opacity,
-        total_count=total_count,
-        items_per_page=items_per_page,
-        db_name=db_name,
-        saved_dbs=saved_dbs,
-        search_history=search_history,
-        current_sort=sort,
-        current_dir=direction,
-        open_tool=tool,
-        select_all_matching=select_all_matching
-    )
-
-@app.route('/fetch_cdx', methods=['POST'])
-def fetch_cdx() -> Response:
-    """Fetch CDX data for a domain and insert new URLs."""
-    domain = request.form.get('domain', '').strip().lower()
-    if not domain:
-        flash("No domain provided for CDX fetch.", "error")
-        return redirect(url_for('index'))
-    if not re.match(r'^(?:[a-zA-Z0-9-]+\.)+[a-zA-Z]{2,63}$', domain):
-        flash("Invalid domain value.", "error")
-        return redirect(url_for('index'))
-    if not _db_loaded():
-        flash("No database loaded.", "error")
-        return redirect(url_for('index'))
-
-    cdx_api = (
-        'http://web.archive.org/cdx/search/cdx'
-        '?url={domain}/*&output=json&fl=original,timestamp,statuscode,mimetype'
-        '&collapse=urlkey&limit=1000'
-    ).format(domain=domain)
-
-    status_mod.push_status('cdx_api_waiting', domain)
-    try:
-        status_mod.push_status('cdx_api_downloading', domain)
-        resp = requests.get(cdx_api, timeout=20)
-        resp.raise_for_status()
-        data = resp.json()
-        status_mod.push_status('cdx_api_download_complete', domain)
-    except Exception as e:
-        flash(f"Error fetching CDX data: {e}", "error")
-        return redirect(url_for('index'))
-
-    inserted = 0
-    for idx, row in enumerate(data):
-        if idx == 0:
-            continue
-        original_url = row[0]
-        timestamp = row[1] if len(row) > 1 else None
-        if len(row) > 2:
-            status_raw = str(row[2])
-            status_code = int(status_raw) if status_raw.isdigit() else None
-        else:
-            status_code = None
-        mime_type = row[3] if len(row) > 3 else None
-        existing = query_db(
-            "SELECT id FROM urls WHERE url = ?",
-            [original_url],
-            one=True
-        )
-        if existing:
-            continue
-        entry_domain = urllib.parse.urlsplit(original_url).hostname or domain
-        execute_db(
-            "INSERT INTO urls (url, domain, timestamp, status_code, mime_type, tags) VALUES (?, ?, ?, ?, ?, ?)",
-            [original_url, entry_domain, timestamp, status_code, mime_type, ""]
-        )
-        inserted += 1
-
-    flash(f"Fetched CDX for {domain}: inserted {inserted} new URLs.", "success")
-    status_mod.push_status('cdx_import_complete', str(inserted))
-    return redirect(url_for('index'))
-
-def _background_import(file_content: bytes) -> None:
-    """Background thread handler for JSON/line-delimited imports."""
-    app_utils._background_import(file_content, app.config['DATABASE'], IMPORT_PROGRESS_FILE)
-
-@app.route('/import_file', methods=['POST'])
-@app.route('/import_json', methods=['POST'])
-def import_file() -> Response:
-    """Import a JSON list of records into the current database."""
-    file = (
-        request.files.get('import_file')
-        or request.files.get('json_file')
-    )
-    if not file:
-        flash("No file uploaded for import.", "error")
-        return redirect(url_for('index'))
-
-    filename = file.filename or ''
-    ext = filename.rsplit('.', 1)[-1].lower()
-
-    if ext != 'json':
-        flash('Please upload a JSON file.', 'error')
-        return redirect(url_for('index'))
-
-    if not _db_loaded():
-        flash('No database loaded.', 'error')
-        return redirect(url_for('index'))
-
-    clear_import_progress()
-    file_content = file.read()
-    set_import_progress('starting', 'Starting import...', 0, 0)
-    thread = threading.Thread(target=_background_import, args=(file_content,))
-    thread.start()
-    flash('Import started! Progress will be shown below.', 'success')
-    return redirect(url_for('index'))
-
-@app.route('/import_progress', methods=['GET'])
-def import_progress() -> Response:
-    """Return JSON describing the current import progress."""
-    prog = get_import_progress()
-    if request.args.get('clear') == '1' and prog.get('status') in ('done', 'failed'):
-        clear_import_progress()
-    # Always supply progress and total, and message for UI
-    return jsonify({
-        'status': prog.get('status', 'idle'),
-        'progress': prog.get('current', 0),
-        'total': prog.get('total', 0),
-        'detail': prog.get('message', '')
-    })
-
-
-@app.route('/status', methods=['GET'])
-def status_route() -> Response:
-    """Return the most recent status event and clear the queue."""
-    evt = status_mod.pop_status()
-    last = evt
-    while True:
-        nxt = status_mod.pop_status()
-        if not nxt:
-            break
-        last = nxt
-    if not last:
-        return ('', 204)
-    return jsonify({'code': last[0], 'message': last[1]})
 
 
 
@@ -474,6 +125,9 @@ from retrorecon.routes import (
     swagger_bp,
     overview_bp,
     tags_bp,
+    assets_bp,
+    import_bp,
+    status_bp,
 )
 app.register_blueprint(notes_bp)
 app.register_blueprint(tools_bp)
@@ -489,14 +143,12 @@ app.register_blueprint(urls_bp)
 app.register_blueprint(swagger_bp)
 app.register_blueprint(overview_bp)
 app.register_blueprint(tags_bp)
+app.register_blueprint(assets_bp)
+app.register_blueprint(import_bp)
+app.register_blueprint(status_bp)
+app.after_request(middleware.add_no_cache_headers)
 
 
-@app.after_request
-def add_no_cache_headers(response: Response) -> Response:
-    """Disable caching so templates always reload."""
-    response.headers['Cache-Control'] = 'no-store'
-    response.headers['Pragma'] = 'no-cache'
-    return response
 
 if __name__ == '__main__':
     if env_db and app.config.get('DATABASE'):
